Fix transformers 5.0 compatibility and add accelerate dependency
- Change torch_dtype to dtype (deprecated in transformers 5.0)
- Add accelerate to dependencies (required for device_map)
- Fix test_entropy_specific_layer test expectation

+++ src/sharingan/core/analyzer.py
@@ -59,7 +59,7 @@ class Sharingan:
         self.model = AutoModelForCausalLM.from_pretrained(
             self.model_name,
             cache_dir=self.config.cache_dir,
-            torch_dtype=self.dtype,
+            dtype=self.dtype,  # transformers 5.0+ uses dtype instead of torch_dtype
             device_map=self.device if self.device != "cpu" else None,
             attn_implementation="eager",  # Required for attention weights
             trust_remote_code=True,
